+++ test/levels.test.js
@@ -2,6 +2,8 @@ const sinon = require('sinon');
 const log = require('../lib');
 const { levels } = require('../lib/config');
 
+const v4 = /^[0-9A-F]{8}-[0-9A-F]{4}-4[0-9A-F]{3}-[89AB][0-9A-F]{3}-[0-9A-F]{12}$|/i;
+
 describe('log levels:', () => {
   let sandbox;
   let sandbox2;
@@ -63,7 +65,7 @@ describe('log levels:', () => {
         timestamp: new Date().toISOString(),
         type: ['technical'],
         level: 'INFO',
-        correlationId: '',
+        correlationId: sinon.match(v4),
       });
     });
   });
@@ -80,7 +82,7 @@ describe('log levels:', () => {
         timestamp: new Date().toISOString(),
         type: ['technical'],
         level: 'INFO',
-        correlationId: '',
+        correlationId: sinon.match(v4),
       });
     });
   });
@@ -97,7 +99,7 @@ describe('log levels:', () => {
         timestamp: new Date().toISOString(),
         type: ['technical'],
         level: 'ERROR',
-        correlationId: '',
+        correlationId: sinon.match(v4),
       });
     });
   });
@@ -114,7 +116,7 @@ describe('log levels:', () => {
         timestamp: new Date().toISOString(),
         type: ['technical'],
         level: 'WARN',
-        correlationId: '',
+        correlationId: sinon.match(v4),
       });
     });
   });
@@ -131,7 +133,7 @@ describe('log levels:', () => {
         timestamp: new Date().toISOString(),
         type: ['technical'],
         level: 'INFO',
-        correlationId: '',
+        correlationId: sinon.match(v4),
       });
     });
   });
@@ -148,7 +150,7 @@ describe('log levels:', () => {
         timestamp: new Date().toISOString(),
         type: ['technical'],
         level: 'DEBUG',
-        correlationId: '',
+        correlationId: sinon.match(v4),
       });
     });
   });