+++ test/requestlog.test.js
@@ -4,6 +4,8 @@ const { levels } = require('../lib/config');
 const log = require('../lib');
 chai.use(require('chai-json-schema'));
 
+const v4 = /^[0-9A-F]{8}-[0-9A-F]{4}-4[0-9A-F]{3}-[89AB][0-9A-F]{3}-[0-9A-F]{12}$|/i;
+
 describe('Logs: -> ', () => {
   let sandbox;
   let sandbox2;
@@ -45,7 +47,7 @@ describe('Logs: -> ', () => {
       timestamp: new Date().toISOString(),
       type: ['technical'],
       level: 'INFO',
-      correlationId: '',
+      correlationId: sinon.match(v4),
       message: '1 2 3',
     });
   });